Rename JSX internal types file
